Improve processing throughput by using a separate handler thread
Start a separate handler thread for interpreting HBase packets and
advancing the internal state. The main thread passes raw packets to the
handler thread via core.async channel.

Note that since the packet stream is stateful, we have to be careful if
we want to further increase the parallelism. It would be possible to
multiplex the packets by the hash value of client address of each
packet. However, using more threads means hbase-packet will occupy more
CPU resource of the server, and it may not be desirable.